+++ app/index.js
@@ -1,18 +1,24 @@
 'use strict';
 
-var yeoman = require( 'yeoman-generator' );
-var yosay = require( 'yosay' );
-var path = require( 'path' );
-var chalk = require( 'chalk' );
-var banner = require( './banner' );
-var request = require( 'request-promise' );
-var filterComponents = require( './lib/filter-components' );
-var updateNotifier = require( './lib/notifier' );
-var fs = require( 'fs' );
-var rimraf = require( 'rimraf' );
-var RegClient = require( 'silent-npm-registry-client' );
-var client = new RegClient();
-var promisify = require( 'promisify-node' );
+const banner = require( './banner' );
+const chalk = require( 'chalk' );
+const download = require( 'download' );
+const filterComponents = require( './lib/filter-components' );
+const fs = require( 'fs' );
+const mkdirp = require( 'mkdirp' );
+const path = require( 'path' );
+const requestPromise = require( 'request-promise' );
+const rimraf = require( 'rimraf' );
+const RegClient = require( 'silent-npm-registry-client' );
+const promisify = require( 'promisify-node' );
+const underscoreString = require( 'underscore.string' );
+const underscoreSlugify = require( 'underscore.string/slugify' );
+const underscore = require( 'underscore' );
+const updateNotifier = require( './lib/notifier' );
+const YeomanGenerator = require( 'yeoman-generator' );
+const yosay = require( 'yosay' );
+
+const client = new RegClient();
 
 // Give the npm registry client a promise interface.
 promisify( client );
@@ -22,27 +28,27 @@ updateNotifier();
 
 // Set list of repos from which we grab standard files.
 // These will be used in the downloadTemplate function in the writing priority.
-var osLibraries = [
+const osLibraries = [
   'https://github.com/cfpb/open-source-project-template/archive/master.zip',
   'https://github.com/cfpb/development/archive/master.zip'
 ];
 
 // Grab a list of all CF components, we'll use it later.
-var components = request( {
+const components = requestPromise( {
   uri:     'https://api.github.com/repos/cfpb/capital-framework/contents/src',
   json:    true,
-  headers: {'user-agent': 'generator-cf'}
+  headers: { 'user-agent': 'generator-cf' }
 } ).then( filterComponents ).catch( console.error );
 
-var manifestCheck = function() {
+const manifestCheck = function() {
   try {
-    var manifest = require( path.join( process.cwd(), 'package.json' ) );
+    const manifest = require( path.join( process.cwd(), 'package.json' ) );
 
     return manifest;
   } catch(e) {};
 };
 
-var CapitalFrameworkGenerator = yeoman.generators.Base.extend( {
+const CapitalFrameworkGenerator = YeomanGenerator.extend( {
 
   initializing: {
 
@@ -63,27 +69,21 @@ var CapitalFrameworkGenerator = yeoman.generators.Base.extend( {
   prompting: {
 
     askForName: function() {
-      var done = this.async();
-
-      this.prompt( {
+      return this.prompt( [ {
         name: 'name',
         message: 'What is the name of your project?',
-        default: this._.humanize( this.existing && this.existing.name || path.basename(process.cwd()) ),
-      }, function( answers ) {
+        default: underscoreString.humanize( this.existing && this.existing.name || path.basename(process.cwd()) ),
+      } ] ).then( function( answers ) {
         this.humanName = answers.name;
-        this.slugname = this._.slugify( answers.name );
-        this.safeSlugname = this.slugname.replace( /-+([a-zA-Z0-9])/g, function ( g ) {
+        this.slugname = underscoreSlugify( answers.name );
+        this.safeSlugname = this.slugname.replace( /-+([a-zA-Z0-9])/g, function( g ) {
             return ' ' + g[1].toUpperCase();
-          }
-        );
-        done();
+        } );
       }.bind( this ) );
     },
 
     askForDescription: function() {
-      var done = this.async();
-
-      var prompts = [{
+      const prompts = [{
         name: 'description',
         message: 'Project\'s description',
         default: this.existing && this.existing.description || 'The best website ever.'
@@ -116,18 +116,17 @@ var CapitalFrameworkGenerator = yeoman.generators.Base.extend( {
         message: 'Author\'s homepage',
         default: this.existing && this.existing.author && this.existing.author.url || 'https://cfpb.github.io/'
       }];
-      this.prompt( prompts, function( answers ) {
+
+      return this.prompt( prompts ).then( function( answers ) {
         this.isCfpbEmployee = answers.authorEmail.indexOf( 'cfpb.gov' ) > -1;
         this.currentYear = ( new Date() ).getFullYear();
         this.props = answers;
-        done();
       }.bind( this ) );
     },
 
     askAboutComponents: function() {
-      var done = this.async();
-      components.then( function( components ) {
-        this.prompt( {
+      return components.then( function( components ) {
+        return this.prompt( [{
           type:    'checkbox',
           name:    'components',
           message: 'You’re about to install CF Core. Which additional CF components would you like in your app?',
@@ -135,7 +134,7 @@ var CapitalFrameworkGenerator = yeoman.generators.Base.extend( {
           default: components.map( function( c ) {
             return c.value;
           } )
-        }, function( answers ) {
+        }] ).then( function( answers ) {
           var versionedComponents = this.components = [];
 
           // Get latest repo tag
@@ -151,32 +150,28 @@ var CapitalFrameworkGenerator = yeoman.generators.Base.extend( {
           answers.components.forEach( function( el ) {
             getLatest( el );
           } );
-          done();
         }.bind( this ) );
       }.bind( this ) );
     },
 
     askForBuildTool: function() {
-      var done = this.async();
-
-      this.prompt( {
+      return this.prompt( [{
         required: true,
         type:     'list',
         name:     'buildToolChoice',
         message:  'Which build tool do you prefer?',
         choices:  [
           {
-            name:  'Grunt',
-            value: 'grunt'
-          },
-          {
             name:  'Gulp',
             value: 'gulp'
+          },
+          {
+            name:  'Grunt',
+            value: 'grunt'
           }
         ]
-      }, function( answers ) {
+      }] ).then( function( answers ) {
         this.buildToolChoice = answers.buildToolChoice;
-        done();
       }.bind( this ) );
     }
   },
@@ -184,12 +179,22 @@ var CapitalFrameworkGenerator = yeoman.generators.Base.extend( {
   writing: {
 
     downloadTemplate: function() {
-      var numLibraries = osLibraries.length;
-      var done = this._.after( numLibraries, this.async() );
-
+      const numLibraries = osLibraries.length;
+      let count = 0;
+      const done = this.async();
       osLibraries.forEach( function( library ) {
-        this.extract( library, '_cache', done );
-      }.bind( this ) );
+        download( library, '_cache', { extract: true } )
+          .then( () => {
+            count++;
+            console.log( 'Finished downloading…', library );
+            if ( numLibraries === count ) {
+              done();
+            }
+          } )
+          .catch( ( err ) => {
+            console.log( 'Error downloading!', library, err );
+          } );
+      } );
     },
 
     appFiles: function() {
@@ -206,7 +211,7 @@ var CapitalFrameworkGenerator = yeoman.generators.Base.extend( {
       files.forEach( function _copy( file ) {
         fs.createReadStream( rootDir + file )
           .pipe( fs.createWriteStream( file ) );
-      }.bind( this ) );
+      } );
 
       // Copy files from the front-end repo.
       var feFiles = ['.eslintrc'];
@@ -214,72 +219,61 @@ var CapitalFrameworkGenerator = yeoman.generators.Base.extend( {
         fs.createReadStream( this.destinationRoot() + '/_cache/development-master/' + file )
           .pipe( fs.createWriteStream( file ) );
       }.bind( this ) );
-
       if ( this.buildToolChoice === 'gulp' ) {
-        this.template( 'gulp/_package.json', 'package.json' );
-        this.template( 'gulp/_README.md', 'README.md' );
-        this.copy( 'gulp/_gulpfile.js', 'gulpfile.js' );
-        this.copy( 'gulp/_setup.sh', 'setup.sh' );
-        this.directory( 'gulp/gulp', 'gulp' );
-        this.directory( 'gulp/config', 'config' );
+        this.fs.copyTpl( this.templatePath( 'gulp/_package.json' ), 'package.json', this );
+        this.fs.copyTpl( this.templatePath( 'gulp/_README.md' ), 'README.md', this );
+        this.fs.copy( this.templatePath( 'gulp/_gulpfile.js' ), 'gulpfile.js' );
+        this.fs.copy( this.templatePath( 'gulp/_setup.sh' ), 'setup.sh' );
+        this.fs.copy( this.templatePath( 'gulp/gulp/' ), 'gulp/' );
+        this.fs.copyTpl( this.templatePath( 'gulp/gulp/config.js' ), 'gulp/config.js', this );
       } else {
-        this.template( 'grunt/_package.json', 'package.json' );
-        this.template( 'grunt/_README.md', 'README.md' );
-        this.copy( 'grunt/_Gruntfile.js', 'Gruntfile.js' );
-        this.copy( 'grunt/_setup.sh', 'setup.sh' );
+        this.fs.copyTpl( this.templatePath( 'grunt/_package.json' ), 'package.json', this );
+        this.fs.copyTpl( this.templatePath( 'grunt/_README.md' ), 'README.md', this );
+        this.fs.copyTpl( this.templatePath( 'grunt/_Gruntfile.js' ), 'Gruntfile.js', this );
+        this.fs.copy( this.templatePath( 'grunt/_setup.sh' ), 'setup.sh' );
       }
     },
 
     srcFiles: function() {
-      this.mkdir( 'src' );
-      this.directory( 'src/static', 'src/static' );
-      this.template( 'src/index.html', 'src/index.html' );
-      this.mkdir( 'dist' );
+      mkdirp( 'src' );
+      this.fs.copyTpl( this.templatePath( 'src/static' ), 'src/static', this );
+      this.fs.copyTpl( this.templatePath( 'src/index.html' ), 'src/index.html', this );
+      mkdirp( 'dist' );
     },
 
     processGitIgnore: function() {
       var fileName = this.destinationRoot() +
                      '/_cache/open-source-project-template-master/.gitignore';
-      var gitignore = this.readFileAsString( fileName );
+      var gitignore = this.fs.read( fileName );
       var done = this.async();
       // Add src/vendor to the end.
       gitignore = gitignore + '\r\nsrc/vendor/';
-      this.writeFileFromString( gitignore, '.gitignore' );
+      this.fs.write( '.gitignore', gitignore );
       // Kill the _cache dir.
       rimraf( this.destinationRoot() + '/_cache', done );
     },
 
     copyNpmrc: function() {
-      this.template( '_npmrc', '.npmrc' );
+      this.templatePath( '_npmrc', '.npmrc' );
     }
-
   },
 
   install: {
-
     installComponents: function() {
-
       if ( this.options['skip-install'] ) return;
-
-      var done = this.async();
-
+      const done = this.async();
       this.spawnCommand( './setup.sh' ).on( 'close', done );
-
     }
-
   },
 
   end: {
 
     bye: function() {
-      var msg = 'All done! Edit the files in the src directory and then ' +
-                '`grunt build` to compile everything into the dist directory.';
+      const msg = `All done! Edit the files in the src directory and then
+                \`${this.buildToolChoice } build\` to compile everything into the dist directory.`;
       this.log( yosay( msg ) );
     }
-
   }
-
-
 } );
 
 module.exports = CapitalFrameworkGenerator;