+++ app/templates/gulp/_gulpfile.js
@@ -12,7 +12,7 @@
   when you run `gulp`.
 */
 
-var requireDir = require( 'require-dir' );
+const requireDir = require( 'require-dir' );
 
 // Require all tasks in gulp/tasks, including subfolders
 requireDir( './gulp/tasks', { recurse: true } );