+++ app/templates/gulp/gulp/tasks/clean.js
@@ -1,9 +1,7 @@
 'use strict';
 
-var gulp = require( 'gulp' );
-var del = require( 'del' );
-var configClean = require( '../config' ).clean;
+const gulp = require( 'gulp' );
+const del = require( 'del' );
+const configClean = require( '../config' ).clean;
 
-gulp.task( 'clean', function() {
-  del( configClean.dest );
-} );
+gulp.task( 'clean', () => del( configClean.dest ) );