+++ app/templates/gulp/gulp/tasks/images.js
@@ -1,15 +1,16 @@
 'use strict';
 
-var gulp = require( 'gulp' );
-var plugins = require( 'gulp-load-plugins' )();
-var configImages = require( '../config' ).images;
-var handleErrors = require( '../utils/handle-errors' );
-var browserSync = require( 'browser-sync' );
+const browserSync = require( 'browser-sync' );
+const configImages = require( '../config' ).images;
+const gulp = require( 'gulp' );
+const gulpChanged = require( 'gulp-changed' );
+const gulpImagemin = require( 'gulp-imagemin' );
+const handleErrors = require( '../utils/handle-errors' );
 
-gulp.task( 'images', function() {
-  return gulp.src( configImages.src )
-    .pipe( plugins.changed( configImages.dest ) )
-    .pipe( plugins.imagemin() )
+gulp.task( 'images', () => {
+  gulp.src( configImages.src )
+    .pipe( gulpChanged( configImages.dest ) )
+    .pipe( gulpImagemin() )
     .on( 'error', handleErrors )
     .pipe( gulp.dest( configImages.dest ) )
     .pipe( browserSync.reload( {